test(positions): split PositionsViewInput chart visibility tests
PositionsViewInputTests was 251 lines after Tasks 1-2 added the new
showsChart / hasHistoricalSeries cases, one line over SwiftLint's
type_body_length budget. Hoist the chart-visibility tests (showsChart,
showsAggregateChart, hasHistoricalSeries) into their own
PositionsViewInputChartTests suite. The remaining file covers totalValue,
totalGainLoss, showsPLPill, and showsGroupSubtotals. No assertion logic
changes.
