+++ MoolahTests/Domain/PositionsViewInputTests.swift
@@ -94,58 +94,6 @@ struct PositionsViewInputTests {
     #expect(!input.showsPLPill)
   }
 
-  @Test("showsChart is false when historicalValue is nil")
-  func chartHiddenWithoutSeries() {
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud,
-      positions: [
-        ValuedPosition(
-          instrument: bhp, quantity: 1, unitPrice: nil,
-          costBasis: amount(50), value: amount(60))
-      ],
-      historicalValue: nil
-    )
-    #expect(!input.showsChart)
-  }
-
-  @Test("showsChart is false when no row carries cost basis")
-  func chartHiddenWithoutAnyCostBasis() {
-    let point = HistoricalValueSeries.Point(
-      date: fixedTestDate, value: 60, cost: 50, contributions: 50)
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud,
-      positions: [
-        ValuedPosition(
-          instrument: bhp, quantity: 1, unitPrice: nil,
-          costBasis: nil, value: amount(100))
-      ],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [point], perInstrument: [:])
-    )
-    #expect(!input.showsChart)
-  }
-
-  @Test("showsAggregateChart is false when any row's value is nil")
-  func aggregateChartHiddenOnFailure() {
-    let point = HistoricalValueSeries.Point(
-      date: fixedTestDate, value: 60, cost: 50, contributions: 50)
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud,
-      positions: [
-        ValuedPosition(
-          instrument: bhp, quantity: 1, unitPrice: nil,
-          costBasis: amount(50), value: amount(60)),
-        ValuedPosition(
-          instrument: bhp, quantity: 1, unitPrice: nil,
-          costBasis: amount(40), value: nil),
-      ],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [point], perInstrument: [:])
-    )
-    #expect(!input.showsAggregateChart)
-    #expect(input.showsChart)  // chart can still render for working instruments
-  }
-
   @Test("totalValue is zero (not nil) for empty positions")
   func totalValueEmptyPositions() {
     let input = PositionsViewInput(
@@ -168,24 +116,6 @@ struct PositionsViewInputTests {
       historicalValue: nil
     )
     #expect(input.showsPLPill)
-  }
-
-  @Test(
-    "showsChart is true when historicalValue has points and at least one row carries cost basis")
-  func chartVisibleWithSeriesAndCostBasis() {
-    let point = HistoricalValueSeries.Point(
-      date: fixedTestDate, value: 60, cost: 50, contributions: 50)
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud,
-      positions: [
-        ValuedPosition(
-          instrument: bhp, quantity: 1, unitPrice: nil,
-          costBasis: amount(50), value: amount(60))
-      ],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [point], perInstrument: [:])
-    )
-    #expect(input.showsChart)
   }
 
   @Test("showsGroupSubtotals is true only when more than one kind is present")
@@ -214,67 +144,6 @@ struct PositionsViewInputTests {
       historicalValue: nil
     )
     #expect(mixed.showsGroupSubtotals)
-  }
-
-  @Test("hasHistoricalSeries is false when historicalValue is nil")
-  func historicalSeriesAbsentWhenNoSeries() {
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud, positions: [], historicalValue: nil)
-    #expect(!input.hasHistoricalSeries)
-  }
-
-  @Test("hasHistoricalSeries is false when total is empty")
-  func historicalSeriesAbsentWhenTotalEmpty() {
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud, positions: [],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [], perInstrument: [:]))
-    #expect(!input.hasHistoricalSeries)
-  }
-
-  @Test("hasHistoricalSeries is true when total has at least one point")
-  func historicalSeriesPresentWhenTotalHasPoints() {
-    let point = HistoricalValueSeries.Point(
-      date: fixedTestDate, value: 100, cost: 80, contributions: 80)
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud, positions: [],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [point], perInstrument: [:]))
-    #expect(input.hasHistoricalSeries)
-  }
-
-  @Test("showsChart is true when positions is empty but historical total has points")
-  func chartVisibleForEmptyPositionsWithHistory() {
-    let point = HistoricalValueSeries.Point(
-      date: fixedTestDate, value: 100, cost: 80, contributions: 80)
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud, positions: [],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [point], perInstrument: [:]))
-    #expect(input.showsChart)
-  }
-
-  @Test("showsChart is false when positions is empty and historical total is empty")
-  func chartHiddenForEmptyPositionsWithoutHistory() {
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud, positions: [],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [], perInstrument: [:]))
-    #expect(!input.showsChart)
-  }
-
-  @Test("showsChart is false when positions has cost basis but historical total is empty")
-  func chartHiddenWhenCostBasisButNoHistoricalPoints() {
-    let input = PositionsViewInput(
-      title: "x", hostCurrency: aud,
-      positions: [
-        ValuedPosition(
-          instrument: bhp, quantity: 1, unitPrice: nil,
-          costBasis: amount(50), value: amount(60))
-      ],
-      historicalValue: HistoricalValueSeries(
-        hostCurrency: aud, total: [], perInstrument: [:]))
-    #expect(!input.showsChart)
   }
 
 }
